refactor: move chunk tests to tests/chunks dir
