+++ tests/chunks/chunk_test.c
@@ -1,4 +1,4 @@
-#include "../inc/tests.h"
+#include "../../inc/tests.h"
 
 void create_top_chunk_test(void** state) {
   t_pagemap* pagemap = (t_pagemap*)*state;
